Added full route domain support
The pool members now include the route domain whether the user
specified it or not (if not, the default route domain is used).
This allows the node names that get created to include the route
domain ID.  By doing this, we won't run into conflicts when the
partition's default route domain is changed.  The fix also
solves issues for pool members with multiple route domains for
a single address on pool members.

There are the following major changes in this commit:
  1. Removed the separate definition for route domains and made it
     be part of the address to mimic BigIP (this would be a breaking
     change, but none of the controller make use of it at this time).
  2. Force route domain to be part of pool addresses when we create
     them (on retrieval from Big-IP add it if missing).
  3. To smooth transition from previous version of cccl, added a
     pre-configure check to see if node names are using the old format.
     If so, the resources are removed before applying the new config
     (this check only occurs at startup).

Other minor changes:
  1. Created the IcrNode/ApiNode class to handle expected naming
     naming convention for nodes.
  2. Prevent pool members from being named. Names are auto-generated and
     include the route domain.