Allow resources within a partition to be ignored
Problem: To have feature parity with the original OpenShift
Big-IP router, we need to allow the user the ability to create
resources and then have the controller manage them.

Solution: Phase one is to build into CCCL the ability to ignore
a whitelist of resources.  This will be done by using the metadata
field within an LTM resource and define a 'cccl-whitelist' attribute.
If the value of the attribute is set to true, then resource will be
ignored, otherwise it will be managed by the controller.