Added merge feature for whitelisted resources
Problem: Currently whitelisted resources are hands off in that the
controller will simply ignore them. We need to have the capability
to modify these resources.

Solution: Added a merge feature that will take the existing
configuration and merge in additional settings from the controller
(but will not remove any settings, even if they conflict).