Enable whitelist of additional LTM resources
Problem: User may want to create other resources (in addition
to the virtual servers) within a controller owned partition that
are whitelisted.

Solution: Pass the properties list of each resource that is allowed
to be whitelisted to the resource base class. The base class will
look for the metatadata field (only resources that support the metadata
field can be whitelisted).