iApps fail to update when user has manager role.
Problem: We currently issue a PUT call for updating iApp resources.
This results in the python SDK making an API call that includes all
attributes, including traffic groups.  Unfortunately, the Big-IP does
not allow traffic groups to be specified by a user with only a Manager
role.

Solution: For just the iApps, switch to using a PATCH call.  In this
case the python SDK will only specify the attributes supplied by CCCL.